Created React version of InstaArt front end.
Made a new React version of the old Django template version of InstaArt. There are still a few loose ends to work out, including pagination and alternating the side that the photo appears on for each piece. Also, there is only one dummy piece. The pieces should be loaded from the back end.