Expandable `TextBox` component
Since some of the descriptions are long, I've wrapped the info cards in a new expandable `TextBox` component. This allows viewers to click a "Read More" button to expand the card's content.

In the future, I'd like this button to only appear if the card goes beyond a certain height. The button shouldn't appear for cards with short descriptions.

Also, I think the button is a but ugly. The button currently overlaps the text as it fades to a pure white background. It might be better to put the button below the fading part, so that it doesn't overlap.